Move `_types.py` to `interactions/_types.py`
